Separate proxy request classification from error handling
Split the proxy handler's top-level utility cluster into request-classification and error-handling modules so the main handler can focus on proxy context assembly while preserving the existing handler export surface for downstream tests and callers.

Constraint: Existing proxy tests and callers still import createProxyHandler, injectContextHeaders, and INTERNAL_PROXY_HEADERS from src/proxy/handler.ts
Rejected: Rework proxy flow and helper boundaries in the same pass | would broaden scope beyond the requested classification/error-handling split
Confidence: medium
Scope-risk: narrow
Reversibility: clean
Directive: Keep request-shape classification helpers and redirect/error parsing helpers outside handler.ts; preserve handler.ts as the compatibility entrypoint unless imports are intentionally migrated later
Tested: deno fmt src/proxy/handler.ts src/proxy/request-classification.ts src/proxy/error-handling.ts; VF_DISABLE_LRU_INTERVAL=1 SSR_TRANSFORM_PER_PROJECT_LIMIT=0 REVALIDATION_PER_PROJECT_LIMIT=0 NODE_ENV=production LOG_FORMAT=text deno test --no-check --allow-all src/proxy/handler.test.ts src/proxy/token-priority.test.ts src/proxy/mode-parity.test.ts; deno check src/proxy/handler.ts src/proxy/request-classification.ts src/proxy/error-handling.ts
Not-tested: Full repository deno test suite outside focused proxy coverage
